yt-dlp-worker/requirements.txt
google-cloud-pubsub==2.19.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.19.0
yt-dlp==2023.7.6
python-dotenv==0.21.0
